Export schematic and Qucs schematic for canonical filters
In this commit they were added to buttons to the user interface: one for downloading the schematic as a SVG image and another for downloading the filter as a Qucs schematic. So far, this option only works for conventional LC filters.

Some research was needed for the download mechanism since I don't have much web programming background. In the end, I've managed to prepare the required files for the download just when the user requires them. This solution uses global variables in the view function. I don't like them a lot, but it is the only thing it came to my mind.